improve cucumber preferences handling
